Move metadata api code to new manager
